Refactor: Organize project structure for better management
- Created src, scripts, config, notebooks, and module_test_outputs directories.
- Moved core Python modules into the src directory.
- Updated import statements to use relative imports within the src package.
- Added __init__.py to src to define it as a package.
- Adjusted output paths for module-level example scripts to save into module_test_outputs/.
- Updated .gitignore to include new artifact/cache directories and common files.
- Tested the changes by running the main workflow, ensuring imports and path resolutions are correct.
