+++ performance_evaluation_report/all_experimental_run_metrics.json
@@ -1,60 +1,31 @@
 {
-    "alpha_0.005_diff_False_lb_60_yrs_3": {
-        "preprocessing_time_seconds": 0.9466080665588379,
-        "selected_features_count": 27,
-        "selected_features_names": "High_^AEX, Low_^AEX, Volume_^AEX, DayOfWeek, Month, SMA_20, SMA_200, EMA_5, MACD_Diff, RSI, BB_Lower, BB_pband, BB_wband, ATR, OBV, Stoch_K, Stoch_D, Williams_R, CCI, ROC, ADX, Volatility_Ret_5D, Volatility_Ret_5D_Raw, Volatility_Ret_10D, Volatility_Ret_10D_Raw, Volatility_Ret_60D, GARCH_Volatility",
-        "used_differencing": false,
-        "plot_directory": "performance_evaluation_report/alpha_0.005_diff_False_lb_60_yrs_3",
-        "model_training_time_seconds": 33.9906747341156,
-        "final_val_loss": 1.4187458753585815,
-        "best_val_loss": 1.4187458753585815,
-        "epochs_trained": 20,
-        "test_set_prediction_time_seconds": 0.9084837436676025,
-        "overall_mse": 1784.1183536751787,
-        "overall_mae": 39.4884933579898,
-        "overall_rmse": 42.23882519288597,
-        "overall_rmse_perc_mean": 4.676748315948168,
-        "overall_mape": 4.358416668135916,
-        "bias_me": 36.533945935837764,
-        "bias_mpe": 3.9943362276192724,
-        "low_vol_mse": 1063.8934683746966,
-        "low_vol_mae": 31.89110819498698,
-        "low_vol_rmse": 32.61737985146411,
-        "mid_vol_mse": 1563.5905070843146,
-        "mid_vol_mae": 37.349606721297555,
-        "mid_vol_rmse": 39.542262290925066,
-        "high_vol_mse": 2927.0216116081506,
-        "high_vol_mae": 51.18541124131943,
-        "high_vol_rmse": 54.101955709642795,
-        "latency_single_pred_ms": 559.4752788543701
-    },
-    "alpha_0.01_diff_False_lb_60_yrs_3": {
-        "preprocessing_time_seconds": 0.1295785903930664,
+    "alpha_0.01_diff_False_lb_60_yrs_2": {
+        "preprocessing_time_seconds": 0.7404747009277344,
         "selected_features_count": 22,
-        "selected_features_names": "High_^AEX, Low_^AEX, Volume_^AEX, Month, SMA_20, SMA_100, SMA_200, EMA_5, MACD_Diff, RSI, BB_Lower, BB_pband, ATR, OBV, Stoch_K, Stoch_D, Williams_R, ROC, ADX, Volatility_Ret_10D, Volatility_Ret_10D_Raw, GARCH_Volatility",
+        "selected_features_names": "High_^AEX, Low_^AEX, Volume_^AEX, DayOfWeek, Month, SMA_20, SMA_100, EMA_5, EMA_200, MACD_Diff, RSI, BB_pband, ATR, OBV, Stoch_K, Stoch_D, Williams_R, ROC, ADX_Pos, Volatility_Ret_10D, Volatility_Ret_10D_Raw, GARCH_Volatility",
         "used_differencing": false,
-        "plot_directory": "performance_evaluation_report/alpha_0.01_diff_False_lb_60_yrs_3",
-        "model_training_time_seconds": 52.54277682304382,
-        "final_val_loss": 1.2749136686325073,
-        "best_val_loss": 1.2749136686325073,
-        "epochs_trained": 20,
-        "test_set_prediction_time_seconds": 0.9081192016601562,
-        "overall_mse": 720.0401074781693,
-        "overall_mae": 23.7166124709109,
-        "overall_rmse": 26.83356307831983,
-        "overall_rmse_perc_mean": 2.9710537725504107,
-        "overall_mape": 2.6258557254538055,
-        "bias_me": 19.339116522606385,
-        "bias_mpe": 2.0899786872183057,
-        "low_vol_mse": 395.90141329438313,
-        "low_vol_mae": 18.373074001736107,
-        "low_vol_rmse": 19.89727150376109,
-        "mid_vol_mse": 554.262692666086,
-        "mid_vol_mae": 21.364172009454258,
-        "mid_vol_rmse": 23.542784301481547,
-        "high_vol_mse": 1361.9188467184485,
-        "high_vol_mae": 33.568995157877595,
-        "high_vol_rmse": 36.90418467760057,
-        "latency_single_pred_ms": 537.3380899429321
+        "plot_directory": "performance_evaluation_report/alpha_0.01_diff_False_lb_60_yrs_2",
+        "model_training_time_seconds": 6.373265743255615,
+        "final_val_loss": 3.313901424407959,
+        "best_val_loss": 3.313901424407959,
+        "epochs_trained": 1,
+        "test_set_prediction_time_seconds": 0.9823541641235352,
+        "overall_mse": 849.5442280492435,
+        "overall_mae": 22.05624796549479,
+        "overall_rmse": 29.146942001679072,
+        "overall_rmse_perc_mean": 3.21614646850463,
+        "overall_mape": 2.502737990931115,
+        "bias_me": -19.531155734592016,
+        "bias_mpe": -2.227738333386078,
+        "low_vol_mse": 251.76176732292654,
+        "low_vol_mae": 13.788982888926636,
+        "low_vol_rmse": 15.867002468107406,
+        "mid_vol_mse": 373.1160814506425,
+        "mid_vol_mae": 16.56805697354403,
+        "mid_vol_rmse": 19.316212916890375,
+        "high_vol_mse": 2358.754447485928,
+        "high_vol_mae": 40.8226610266644,
+        "high_vol_rmse": 48.567009867665604,
+        "latency_single_pred_ms": 534.8434925079346
     }
 }
