+++ performance_evaluation_report/all_experimental_run_summary.csv
@@ -1,3 +1,2 @@
 run_id,lasso_alpha,differencing,look_back,years_data,preprocessing_time_seconds,selected_features_count,selected_features_names,used_differencing,plot_directory,model_training_time_seconds,final_val_loss,best_val_loss,epochs_trained,test_set_prediction_time_seconds,overall_mse,overall_mae,overall_rmse,overall_rmse_perc_mean,overall_mape,bias_me,bias_mpe,low_vol_mse,low_vol_mae,low_vol_rmse,mid_vol_mse,mid_vol_mae,mid_vol_rmse,high_vol_mse,high_vol_mae,high_vol_rmse,latency_single_pred_ms
-alpha_0.005_diff_False_lb_60_yrs_3,0.005,False,60,3,0.9466080665588379,27,"High_^AEX, Low_^AEX, Volume_^AEX, DayOfWeek, Month, SMA_20, SMA_200, EMA_5, MACD_Diff, RSI, BB_Lower, BB_pband, BB_wband, ATR, OBV, Stoch_K, Stoch_D, Williams_R, CCI, ROC, ADX, Volatility_Ret_5D, Volatility_Ret_5D_Raw, Volatility_Ret_10D, Volatility_Ret_10D_Raw, Volatility_Ret_60D, GARCH_Volatility",False,performance_evaluation_report/alpha_0.005_diff_False_lb_60_yrs_3,33.9906747341156,1.4187458753585815,1.4187458753585815,20,0.9084837436676025,1784.1183536751787,39.4884933579898,42.23882519288597,4.676748315948168,4.358416668135916,36.533945935837764,3.9943362276192724,1063.8934683746966,31.89110819498698,32.61737985146411,1563.5905070843146,37.349606721297555,39.542262290925066,2927.0216116081506,51.18541124131943,54.101955709642795,559.4752788543701
-alpha_0.01_diff_False_lb_60_yrs_3,0.01,False,60,3,0.1295785903930664,22,"High_^AEX, Low_^AEX, Volume_^AEX, Month, SMA_20, SMA_100, SMA_200, EMA_5, MACD_Diff, RSI, BB_Lower, BB_pband, ATR, OBV, Stoch_K, Stoch_D, Williams_R, ROC, ADX, Volatility_Ret_10D, Volatility_Ret_10D_Raw, GARCH_Volatility",False,performance_evaluation_report/alpha_0.01_diff_False_lb_60_yrs_3,52.54277682304382,1.2749136686325073,1.2749136686325073,20,0.9081192016601562,720.0401074781693,23.7166124709109,26.83356307831983,2.9710537725504107,2.6258557254538055,19.339116522606385,2.0899786872183057,395.90141329438313,18.373074001736107,19.89727150376109,554.262692666086,21.364172009454258,23.542784301481547,1361.9188467184485,33.568995157877595,36.90418467760057,537.3380899429321
+alpha_0.01_diff_False_lb_60_yrs_2,0.01,False,60,2,0.7404747009277344,22,"High_^AEX, Low_^AEX, Volume_^AEX, DayOfWeek, Month, SMA_20, SMA_100, EMA_5, EMA_200, MACD_Diff, RSI, BB_pband, ATR, OBV, Stoch_K, Stoch_D, Williams_R, ROC, ADX_Pos, Volatility_Ret_10D, Volatility_Ret_10D_Raw, GARCH_Volatility",False,performance_evaluation_report/alpha_0.01_diff_False_lb_60_yrs_2,6.373265743255615,3.313901424407959,3.313901424407959,1,0.9823541641235352,849.5442280492435,22.05624796549479,29.146942001679072,3.21614646850463,2.502737990931115,-19.531155734592016,-2.227738333386078,251.76176732292654,13.788982888926636,15.867002468107406,373.1160814506425,16.56805697354403,19.316212916890375,2358.754447485928,40.8226610266644,48.567009867665604,534.8434925079346
