+++ src/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the src directory a Python package.
